requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
pyyaml>=6.0
pyarrow>=12.0
# Optional — only install if you plan to use pytorch_mlp
# torch>=2.0
